Small changes as follow-up of #591
* Move bugzilla model tests to test_models

* Delete module docstring and pylint ignore in test_runner.py

- if pylint can't enumerate test fixtures, that itself is a clue that
  something is wrong
- The organization of the test directory mirrors that application code
  directory, so it's evident what this module tests

* Update acceptance date of ADR
