+++ docs/adrs/002-action-customizaiton.md
@@ -1,7 +1,7 @@
 # Restrict Workflow Customization and Validate Default Action Parameters
 
 - Status: Accepted
-- Date: 2023-06-23
+- Date: 2023-06-27
 
 Tracking issue: [#544](https://github.com/mozilla/jira-bugzilla-integration/issues/544)
